requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
boto3==1.35.62
boto3-stubs==1.35.62
botocore==1.35.62
botocore-stubs==1.35.62
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain-core==0.3.18
langchain-openai==0.2.8
langsmith==0.1.143
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-boto3-cloudformation==1.35.41
mypy-boto3-dynamodb==1.35.60
mypy-boto3-ec2==1.35.60
mypy-boto3-iam==1.35.61
mypy-boto3-lambda==1.35.58
mypy-boto3-rds==1.35.59
mypy-boto3-s3==1.35.61
mypy-boto3-sqs==1.35.0
narwhals==1.13.5
numpy==2.0.2
openai==1.54.4
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
s3transfer==0.10.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.40.1
tenacity==9.0.0
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
types-awscrt==0.23.0
types-s3transfer==0.10.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
watchdog==6.0.0
